+++ CMakeLists.txt
@@ -80,7 +80,7 @@ else()
 endif()
 
 add_custom_target(models
-  COMMAND ${RBC} -Bcore           core.rb
+  COMMAND ${RBC} -Bfloat_ext      float_ext.rb
   COMMAND ${RBC} -Bkeyboard       keyboard.rb
   COMMAND ${RBC} -Brgb            rgb.rb
   COMMAND ${RBC} -Brotary_encoder rotary_encoder.rb
